Add a special fix for weird challenge 4 behavior
It appears that challenge 4's backend processing script expects
the HTTP referer header to be set to a specific URL. This patches
the generic Enigma HTTP request handler to do the Right Thing (TM).

Naturally, this is not stated in the challenge description and was
found to work by trial and error.